+++ packages/package-json-lint-config-terra/package-json-lint.config.js
@@ -3,6 +3,6 @@ module.exports = {
     'require-no-hard-coded-dependency-versions': { severityType: 'error' },
     'require-no-terra-base-peer-dependency-versions': { severityType: 'warn' },
     'require-theme-context-versions': { severityType: 'warn' },
-    'require-no-unnecessary-dependencies': { severityType: 'warn' },
+    'require-no-unnecessary-dependency': { severityType: 'warn' },
   },
 };